Drop old stable debian release from distributions.
This will stop bloom from attempting to generate Debian Bullseye release
metadata when running releases for these distributions.

This will not prevent past releases from being available in the ros2-gbp
future releases will not include metadata for these platforms.

I am opening one PR for the initial review but I think it would be best
to merge this change distribution-by-distribution after each one has
completed a sync so that there is a tag which represents the last set of
of releases where Debian metadata was attempted. I'll use this PR to
coordinate those changes.